requirements.txt
# ============================================================================
# Core Dependencies
# ============================================================================
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.10.0

# ============================================================================
# Web Scraping
# ============================================================================
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0

# ============================================================================
# Data Processing
# ============================================================================
pandas==2.2.0
pyarrow==15.0.0
fastparquet==2024.2.0

# ============================================================================
# Vector Database
# ============================================================================
chromadb==0.4.22
sentence-transformers==2.3.1

# ============================================================================
# Database
# ============================================================================
psycopg2-binary==2.9.9
redis==5.0.1

# ============================================================================
# Utilities
# ============================================================================
python-dotenv==1.0.1
